feat: Redirect logged in users to the list page
