+++ pages/users.js
@@ -2,6 +2,7 @@ import { useContext } from "react";
 import AppContext from "../context/AppContext";
 import Link from "next/link";
 import RouteGuard from "../guards/RouteGuard";
+import Profile from "../components/Profile";
 
 const UsersPage = ({ levels, milestones }) => {
   const { data, loading } = useContext(AppContext);
@@ -11,13 +12,16 @@ const UsersPage = ({ levels, milestones }) => {
       {loading ? (
         "Loading..."
       ) : (
-        <ul>
-          {Object.keys(data.users).map((userName) => (
-            <li key={userName}>
-              <Link href={`/user/${userName}`}>{userName}</Link>
-            </li>
-          ))}
-        </ul>
+        <div>
+          <Profile />
+          <ul>
+            {Object.keys(data.users).map((userName) => (
+              <li key={userName}>
+                <Link href={`/user/${userName}`}>{userName}</Link>
+              </li>
+            ))}
+          </ul>
+        </div>
       )}
     </RouteGuard>
   );
